Added Copy functionality to Feed page
